requirements.txt
flask
sqlalchemy
flask-httpauth
pyyaml
rich
